requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
cycler==0.10.0
dj-database-url==0.5.0
djangorestframework==3.5.3
gunicorn==20.0.4
heroku==0.1.4
html5lib==1.0.1
idna==2.9
importlib-metadata==1.6.0
kiwisolver==1.2.0
kombu==4.6.8
matplotlib==3.2.1
numpy==1.18.1
opencv-contrib-python==4.2.0.32
opencv-contrib-python-headless==4.2.0.32
opencv-python==4.2.0.32
Pillow==7.0.0
pip3-multiple-versions==0.2.2
pyglet==1.5.5
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2019.3
pyzbar==0.1.8
reportlab==3.5.34
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.9
vine==1.3.0
webencodings==0.5.1
whitenoise==5.1.0
xhtml2pdf==0.2.4
zipp==3.1.0
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
cycler==0.10.0
dhtml5lib==1.0.1
idna==2.9
importlib-metadata==1.6.0
kiwisolver==1.2.0
kombu==4.6.8
matplotlib==3.2.1
numpy==1.18.1
opencv-contrib-python==4.2.0.32
opencv-contrib-python-headless==4.2.0.32
opencv-python==4.2.0.32
Pillow==7.0.0
pip3-multiple-versions==0.2.2
pytz==2019.3
pyzbar==0.1.8
reportlab==3.5.34
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.9
vine==1.3.0
webencodings==0.5.1
whitenoise==5.1.0
xhtml2pdf==0.2.4
zipp==3.1.0
